refactor: migrate BIND tests to pytest infrastructure
- Move BIND test suite from standalone script to pytest-based structure
- Create tests/bind/ directory with proper conftest.py and test files
- Consolidate MockSocksServer class in conftest.py for reuse
- Update Makefile to use pytest format with HTML/JSON reporting
- Remove old test_bind.py script
- All original test functionality preserved with better organization

Testing infrastructure modernization complete.
