--- tests/comprehensive/scripts/test_bind.py
@@ -1,1139 +1,0 @@
-#!/usr/bin/env python3
-"""
-BIND functionality comprehensive test suite for RedProxy
-Tests SOCKS BIND command support across listeners, connectors, and direct connections
-"""
-
-import asyncio
-import socket
-import struct
-import json
-import os
-import signal
-import time
-from typing import Dict, List, Optional, Tuple
-from contextlib import asynccontextmanager
-
-from lib.test_utils import TestLogger, setup_test_environment, wait_for_service
-from lib.test_reporter import TestReporter, TestResult
-
-class Colors:
-    RED = '\033[0;31m'
-    GREEN = '\033[0;32m'
-    YELLOW = '\033[1;33m'
-    BLUE = '\033[0;34m'
-    NC = '\033[0m'
-
-
-class SocksBindTester:
-    """Test SOCKS BIND functionality"""
-    
-    def __init__(self, host: str, port: int):
-        self.host = host
-        self.port = port
-    
-    async def socks5_bind_request(self, target_host: str, target_port: int) -> Tuple[bool, Optional[Tuple[str, int]]]:
-        """
-        Perform SOCKS5 BIND request and return success status and bound address
-        """
-        try:
-            TestLogger.info(f"Attempting SOCKS5 BIND to {target_host}:{target_port}")
-            
-            # Connect to SOCKS5 proxy
-            reader, writer = await asyncio.open_connection(self.host, self.port)
-            
-            try:
-                # SOCKS5 authentication negotiation
-                auth_request = b'\x05\x01\x00'  # Version 5, 1 method, no auth
-                writer.write(auth_request)
-                await writer.drain()
-                
-                auth_response = await reader.read(2)
-                if len(auth_response) != 2 or auth_response != b'\x05\x00':
-                    TestLogger.error(f"SOCKS5 auth failed: {auth_response.hex()}")
-                    return False, None
-                
-                # SOCKS5 BIND request
-                bind_request = bytearray([0x05, 0x02, 0x00])  # Version, BIND command, reserved
-                
-                if target_host.replace('.', '').isdigit():
-                    # IPv4 address
-                    bind_request.append(0x01)  # IPv4 address type
-                    bind_request.extend(socket.inet_aton(target_host))
-                else:
-                    # Domain name
-                    bind_request.append(0x03)  # Domain name type
-                    bind_request.append(len(target_host))
-                    bind_request.extend(target_host.encode())
-                
-                bind_request.extend(struct.pack('>H', target_port))  # Port in network byte order
-                
-                writer.write(bind_request)
-                await writer.drain()
-                
-                # Read BIND response (first response with bound address)
-                response = await reader.read(1024)
-                if len(response) < 6:
-                    TestLogger.error(f"SOCKS5 BIND response too short: {len(response)} bytes")
-                    return False, None
-                
-                if response[0] != 0x05:
-                    TestLogger.error(f"Invalid SOCKS5 version in response: {response[0]}")
-                    return False, None
-                
-                if response[1] != 0x00:
-                    TestLogger.error(f"SOCKS5 BIND failed with error code: {response[1]}")
-                    return False, None
-                
-                # Parse bound address from response
-                addr_type = response[3]
-                if addr_type == 0x01:  # IPv4
-                    if len(response) < 10:
-                        TestLogger.error("SOCKS5 IPv4 response too short")
-                        return False, None
-                    bound_ip = socket.inet_ntoa(response[4:8])
-                    bound_port = struct.unpack('>H', response[8:10])[0]
-                elif addr_type == 0x03:  # Domain name
-                    domain_len = response[4]
-                    if len(response) < 5 + domain_len + 2:
-                        TestLogger.error("SOCKS5 domain response too short")
-                        return False, None
-                    bound_ip = response[5:5+domain_len].decode()
-                    bound_port = struct.unpack('>H', response[5+domain_len:7+domain_len])[0]
-                else:
-                    TestLogger.error(f"Unsupported address type: {addr_type}")
-                    return False, None
-                
-                TestLogger.info(f"SOCKS5 BIND successful, bound to {bound_ip}:{bound_port}")
-                
-                # Keep connection open for potential second response
-                # In real usage, this would wait for incoming connections
-                return True, (bound_ip, bound_port)
-                
-            finally:
-                writer.close()
-                await writer.wait_closed()
-                
-        except Exception as e:
-            TestLogger.error(f"SOCKS5 BIND request failed: {e}")
-            return False, None
-    
-    async def socks4_bind_request(self, target_ip: str, target_port: int) -> Tuple[bool, Optional[Tuple[str, int]]]:
-        """
-        Perform SOCKS4 BIND request and return success status and bound address
-        """
-        try:
-            TestLogger.info(f"Attempting SOCKS4 BIND to {target_ip}:{target_port}")
-            
-            # Connect to SOCKS proxy
-            reader, writer = await asyncio.open_connection(self.host, self.port)
-            
-            try:
-                # SOCKS4 BIND request
-                bind_request = bytearray([0x04, 0x02])  # Version 4, BIND command
-                bind_request.extend(struct.pack('>H', target_port))  # Port
-                bind_request.extend(socket.inet_aton(target_ip))  # IP address
-                bind_request.append(0x00)  # Empty user ID
-                
-                writer.write(bind_request)
-                await writer.drain()
-                
-                # Read BIND response
-                response = await reader.read(8)
-                if len(response) != 8:
-                    TestLogger.error(f"SOCKS4 BIND response wrong length: {len(response)}")
-                    return False, None
-                
-                if response[0] != 0x00:
-                    TestLogger.error(f"Invalid SOCKS4 response byte: {response[0]}")
-                    return False, None
-                
-                if response[1] != 0x5a:  # Request granted
-                    TestLogger.error(f"SOCKS4 BIND failed with code: {response[1]:02x}")
-                    return False, None
-                
-                # Parse bound address
-                bound_port = struct.unpack('>H', response[2:4])[0]
-                bound_ip = socket.inet_ntoa(response[4:8])
-                
-                TestLogger.info(f"SOCKS4 BIND successful, bound to {bound_ip}:{bound_port}")
-                return True, (bound_ip, bound_port)
-                
-            finally:
-                writer.close()
-                await writer.wait_closed()
-                
-        except Exception as e:
-            TestLogger.error(f"SOCKS4 BIND request failed: {e}")
-            return False, None
-
-
-class EchoServer:
-    """Simple echo server for testing connections to BIND ports"""
-    
-    def __init__(self, host: str = "0.0.0.0", port: int = 0):
-        self.host = host
-        self.port = port
-        self.server = None
-        self.actual_port = None
-    
-    async def start(self) -> int:
-        """Start echo server and return actual port"""
-        TestLogger.info(f"Starting echo server on {self.host}:{self.port}")
-        
-        async def handle_client(reader, writer):
-            try:
-                addr = writer.get_extra_info('peername')
-                TestLogger.info(f"Echo server: connection from {addr}")
-                
-                # Simple echo loop
-                while True:
-                    data = await reader.read(1024)
-                    if not data:
-                        break
-                    
-                    TestLogger.info(f"Echo server received: {data.decode()}")
-                    writer.write(data)
-                    await writer.drain()
-                    
-            except Exception as e:
-                TestLogger.warn(f"Echo server client error: {e}")
-            finally:
-                writer.close()
-                await writer.wait_closed()
-        
-        self.server = await asyncio.start_server(handle_client, self.host, self.port)
-        self.actual_port = self.server.sockets[0].getsockname()[1]
-        TestLogger.info(f"Echo server listening on port {self.actual_port}")
-        return self.actual_port
-    
-    async def stop(self):
-        """Stop echo server"""
-        if self.server:
-            self.server.close()
-            await self.server.wait_closed()
-            TestLogger.info("Echo server stopped")
-
-
-class MockSocksServer:
-    """Mock SOCKS server for testing SOCKS connector BIND functionality"""
-    
-    def __init__(self, host: str = "0.0.0.0", port: int = 0):
-        self.host = host
-        self.port = port
-        self.server = None
-        self.actual_port = None
-        self.bind_responses = []
-        self._client_tasks = set()  # Track client handling tasks for proper cleanup
-    
-    async def start(self) -> int:
-        """Start mock SOCKS server"""
-        TestLogger.info(f"Starting mock SOCKS server on {self.host}:{self.port}")
-        
-        async def handle_client(reader, writer):
-            task = None
-            try:
-                # Get current task for cleanup tracking
-                task = asyncio.current_task()
-                self._client_tasks.add(task)
-                
-                await self._handle_socks_session(reader, writer)
-            except Exception as e:
-                TestLogger.error(f"Mock SOCKS server error: {e}")
-            finally:
-                writer.close()
-                await writer.wait_closed()
-                if task and task in self._client_tasks:
-                    self._client_tasks.discard(task)
-        
-        self.server = await asyncio.start_server(handle_client, self.host, self.port)
-        self.actual_port = self.server.sockets[0].getsockname()[1]
-        TestLogger.info(f"Mock SOCKS server listening on port {self.actual_port}")
-        return self.actual_port
-    
-    async def _handle_socks_session(self, reader, writer):
-        """Handle SOCKS protocol session"""
-        # Read version byte first
-        version_data = await reader.read(1)
-        if not version_data:
-            return
-        
-        version = version_data[0]
-        
-        if version == 5:
-            await self._handle_socks5(reader, writer, version_data)
-        elif version == 4:
-            await self._handle_socks4(reader, writer, version_data)
-        else:
-            TestLogger.error(f"Unsupported SOCKS version: {version}")
-    
-    async def _handle_socks5(self, reader, writer, first_byte):
-        """Handle SOCKS5 protocol"""
-        # Read rest of auth request
-        auth_data = await reader.read(1)
-        if not auth_data:
-            return
-        
-        num_methods = auth_data[0]
-        methods = await reader.read(num_methods)
-        
-        # Accept no auth
-        writer.write(b'\x05\x00')
-        await writer.drain()
-        
-        # Read BIND request
-        request = await reader.read(4)  # VER CMD RSV ATYP
-        if len(request) < 4 or request[1] != 0x02:  # Not BIND
-            writer.write(b'\x05\x07' + b'\x00' * 8)  # Command not supported
-            await writer.drain()
-            return
-        
-        addr_type = request[3]
-        if addr_type == 0x01:  # IPv4
-            addr_data = await reader.read(6)  # 4 bytes IP + 2 bytes port
-        elif addr_type == 0x03:  # Domain
-            domain_len = (await reader.read(1))[0]
-            addr_data = await reader.read(domain_len + 2)  # domain + port
-        else:
-            writer.write(b'\x05\x08' + b'\x00' * 8)  # Address type not supported
-            await writer.drain()
-            return
-        
-        # Send successful BIND response with mock bound address
-        response = b'\x05\x00\x00\x01'  # Version, success, reserved, IPv4
-        response += socket.inet_aton('127.0.0.1')  # Mock bound IP
-        response += struct.pack('>H', 8080)  # Mock bound port
-        
-        writer.write(response)
-        await writer.drain()
-        
-        TestLogger.info("Mock SOCKS5 BIND response sent")
-        
-        # Simulate waiting for connection (send second response after delay)
-        await asyncio.sleep(1)
-        
-        # Send second response indicating connection from peer
-        peer_response = b'\x05\x00\x00\x01'  # Version, success, reserved, IPv4
-        peer_response += socket.inet_aton('192.168.1.100')  # Mock peer IP
-        peer_response += struct.pack('>H', 12345)  # Mock peer port
-        
-        writer.write(peer_response)
-        await writer.drain()
-        
-        TestLogger.info("Mock SOCKS5 second BIND response sent (peer connected)")
-    
-    async def _handle_socks4(self, reader, writer, first_byte):
-        """Handle SOCKS4 protocol"""
-        # Read rest of SOCKS4 request
-        request_data = await reader.read(7)  # CMD(1) + PORT(2) + IP(4)
-        if len(request_data) < 7 or request_data[0] != 0x02:  # Not BIND
-            response = b'\x00\x5b' + b'\x00' * 6  # Request rejected
-            writer.write(response)
-            await writer.drain()
-            return
-        
-        # Read user ID (null terminated)
-        while True:
-            byte = await reader.read(1)
-            if not byte or byte == b'\x00':
-                break
-        
-        # Send successful BIND response
-        response = b'\x00\x5a'  # VN, CD (success)
-        response += struct.pack('>H', 8080)  # Mock bound port
-        response += socket.inet_aton('127.0.0.1')  # Mock bound IP
-        
-        writer.write(response)
-        await writer.drain()
-        
-        TestLogger.info("Mock SOCKS4 BIND response sent")
-        
-        # Simulate connection from peer
-        await asyncio.sleep(1)
-        
-        peer_response = b'\x00\x5a'  # VN, CD (success)
-        peer_response += struct.pack('>H', 12345)  # Mock peer port
-        peer_response += socket.inet_aton('192.168.1.100')  # Mock peer IP
-        
-        writer.write(peer_response)
-        await writer.drain()
-        
-        TestLogger.info("Mock SOCKS4 second BIND response sent (peer connected)")
-    
-    async def stop(self):
-        """Stop mock SOCKS server with proper cleanup"""
-        if self.server:
-            TestLogger.info("Stopping mock SOCKS server...")
-            self.server.close()
-            await self.server.wait_closed()
-            
-            # Cancel any remaining client tasks
-            if self._client_tasks:
-                TestLogger.info(f"Cancelling {len(self._client_tasks)} remaining client tasks")
-                for task in self._client_tasks:
-                    if not task.done():
-                        task.cancel()
-                
-                # Wait for cancelled tasks to complete
-                await asyncio.gather(*self._client_tasks, return_exceptions=True)
-                self._client_tasks.clear()
-            
-            TestLogger.info("Mock SOCKS server stopped")
-
-
-class BindTestSuite:
-    """Comprehensive BIND test suite"""
-    
-    def __init__(self, env):
-        self.env = env
-        self.reporter = TestReporter(output_dir="/reports")
-        
-        # Set environment for the report
-        self.reporter.set_environment({
-            "test_type": "bind_functionality",
-            "redproxy_version": os.environ.get("REDPROXY_VERSION", "unknown"),
-            "config_file": "/config/bind-test.yaml"
-        })
-        
-        self.suite = self.reporter.create_suite("BIND Functionality Tests")
-    
-    def add_test_result(self, name: str, success: bool, message: str = "", duration: float = 0.0):
-        """Helper to add test results to the suite"""
-        status = "passed" if success else "failed"
-        error_message = None if success else message
-        self.suite.tests.append(TestResult(
-            name=name,
-            status=status,
-            duration=duration,
-            error_message=error_message
-        ))
-    
-    async def test_socks_listener_bind_support(self) -> bool:
-        """Test SOCKS listener BIND command support with different configurations"""
-        TestLogger.test("Testing SOCKS Listener BIND Support")
-        
-        # Test 1: BIND allowed (port 1081)
-        TestLogger.info("Testing allow_bind=true, enforce_bind_address=false (port 1081)")
-        socks_tester_allowed = SocksBindTester(self.env.redproxy_host, 1081)
-        success_allowed, bound_addr_allowed = await socks_tester_allowed.socks5_bind_request("0.0.0.0", 0)
-        
-        if success_allowed and bound_addr_allowed:
-            TestLogger.info(f"BIND allowed test successful: {bound_addr_allowed[0]}:{bound_addr_allowed[1]}")
-            self.add_test_result("bind_allowed_test", True, f"Bound to {bound_addr_allowed[0]}:{bound_addr_allowed[1]}")
-        else:
-            TestLogger.error("BIND allowed test failed")
-            self.add_test_result("bind_allowed_test", False, "BIND request failed on allow_bind=true")
-        
-        # Test 2: BIND denied (port 1082)
-        TestLogger.info("Testing allow_bind=false (port 1082)")
-        socks_tester_denied = SocksBindTester(self.env.redproxy_host, 1082)
-        success_denied, bound_addr_denied = await socks_tester_denied.socks5_bind_request("0.0.0.0", 0)
-        
-        if not success_denied:
-            TestLogger.info("BIND denied test successful - request properly rejected")
-            self.add_test_result("bind_denied_test", True, "BIND properly denied when allow_bind=false")
-        else:
-            TestLogger.error("BIND denied test failed - request should have been rejected")
-            self.add_test_result("bind_denied_test", False, "BIND was allowed when allow_bind=false")
-        
-        # Test 3: BIND enforce address (port 1083)
-        TestLogger.info("Testing allow_bind=true, enforce_bind_address=true (port 1083)")
-        socks_tester_enforce = SocksBindTester(self.env.redproxy_host, 1083)
-        success_enforce, bound_addr_enforce = await socks_tester_enforce.socks5_bind_request("0.0.0.0", 8888)  # Request specific port on valid IP
-        
-        if success_enforce and bound_addr_enforce:
-            # Should get system-assigned port, not the requested port 8888
-            if bound_addr_enforce[1] != 8888:
-                TestLogger.info(f"BIND enforce test successful - got system port: {bound_addr_enforce[0]}:{bound_addr_enforce[1]} (not requested port 8888)")
-                self.add_test_result("bind_enforce_test", True, f"Address enforced: got port {bound_addr_enforce[1]} instead of requested port 8888")
-            else:
-                TestLogger.error("BIND enforce test failed - got requested port instead of system-assigned")
-                self.add_test_result("bind_enforce_test", False, "Port not enforced - got requested port 8888")
-        else:
-            TestLogger.error("BIND enforce test failed - request failed")
-            self.add_test_result("bind_enforce_test", False, "BIND request failed on enforce_bind_address=true")
-        
-        return success_allowed and not success_denied and success_enforce
-    
-    async def test_socks_connector_bind_functionality(self) -> bool:
-        """Test SOCKS connector BIND functionality with end-to-end integration"""
-        TestLogger.test("Testing SOCKS Connector BIND Integration")
-        
-        # Start mock SOCKS server that will act as upstream
-        mock_server = MockSocksServer()
-        mock_port = await mock_server.start()
-        
-        try:
-            # Give mock server time to start
-            await asyncio.sleep(0.5)
-            
-            # Test 1: Direct connection to mock server (validate mock works)
-            TestLogger.info("Phase 1: Testing mock SOCKS server directly")
-            tester = SocksBindTester("127.0.0.1", mock_port)
-            success, bound_addr = await tester.socks5_bind_request("example.com", 80)
-            
-            if not success or not bound_addr:
-                TestLogger.error("Mock SOCKS server BIND failed")
-                self.add_test_result("socks_connector_bind", False, "Mock server validation failed")
-                return False
-            
-            TestLogger.info(f"Mock SOCKS server BIND successful: {bound_addr}")
-            
-            # Test 2: Test SOCKS4 BIND on mock server
-            TestLogger.info("Phase 2: Testing SOCKS4 BIND on mock server")
-            success_v4, bound_addr_v4 = await tester.socks4_bind_request("192.168.1.100", 8080)
-            
-            if success_v4 and bound_addr_v4:
-                TestLogger.info(f"Mock SOCKS4 BIND successful: {bound_addr_v4}")
-            else:
-                TestLogger.warn("Mock SOCKS4 BIND failed (may be expected)")
-            
-            # Test 3: Test with authentication
-            TestLogger.info("Phase 3: Testing authenticated SOCKS5 BIND")
-            # Note: Our mock server accepts any auth, so this should work
-            success_auth, bound_addr_auth = await tester.socks5_bind_request("example.com", 443)
-            
-            # Evaluate results
-            passed_tests = sum([
-                success and bound_addr is not None,  # Basic SOCKS5 BIND
-                success_v4 and bound_addr_v4 is not None,  # SOCKS4 BIND
-                success_auth and bound_addr_auth is not None,  # Authenticated BIND
-            ])
-            
-            TestLogger.info(f"SOCKS connector integration tests: {passed_tests}/3 passed")
-            
-            if passed_tests >= 2:  # Allow one failure
-                self.add_test_result("socks_connector_bind", True, f"Integration tests: {passed_tests}/3 passed")
-                return True
-            else:
-                self.add_test_result("socks_connector_bind", False, f"Integration tests: only {passed_tests}/3 passed")
-                return False
-                
-        finally:
-            await mock_server.stop()
-    
-    async def test_direct_connector_bind_functionality(self) -> bool:
-        """Test DirectConnector BIND functionality"""
-        TestLogger.test("Testing DirectConnector BIND Functionality")
-        
-        # Direct connector BIND would typically be tested by:
-        # 1. Configuring RedProxy with DirectConnector for BIND
-        # 2. Making a BIND request through SOCKS listener
-        # 3. Verifying that DirectConnector creates a listening socket
-        
-        # For this test, we'll verify that the configuration supports BIND
-        # and that basic socket binding works
-        
-        try:
-            # Test basic socket binding capability
-            test_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            test_socket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-            test_socket.bind(('127.0.0.1', 0))
-            bound_port = test_socket.getsockname()[1]
-            test_socket.close()
-            
-            TestLogger.info(f"Basic socket bind test successful on port {bound_port}")
-            self.add_test_result("direct_connector_bind", True, f"Basic bind successful on port {bound_port}")
-            return True
-            
-        except Exception as e:
-            TestLogger.error(f"Direct connector bind test failed: {e}")
-            self.add_test_result("direct_connector_bind", False, f"Bind failed: {e}")
-            return False
-    
-    async def test_bind_with_echo_server(self) -> bool:
-        """Test BIND functionality with proper client-server connection flow"""
-        TestLogger.test("Testing BIND with Echo Server Connection Flow")
-        
-        try:
-            # Step 1: Create SOCKS5 BIND request (this sets up a listening socket)
-            TestLogger.info("Step 1: Setting up SOCKS5 BIND operation")
-            socks_tester = SocksBindTester(self.env.redproxy_host, 1081)  # Use bind-allowed port
-            success, bound_addr = await socks_tester.socks5_bind_request("0.0.0.0", 0)
-            
-            if not success or not bound_addr:
-                self.add_test_result("bind_echo_test", False, "SOCKS5 BIND setup failed")
-                return False
-            
-            TestLogger.info(f"BIND setup successful, listening on {bound_addr[0]}:{bound_addr[1]}")
-            
-            # Step 2: Start a client that will connect to the BIND address
-            # This simulates what FTP data connections or similar protocols would do
-            TestLogger.info("Step 2: Simulating client connecting to BIND address")
-            
-            # Create a simple client task that will connect to the bound address
-            async def simulate_client_connection():
-                try:
-                    await asyncio.sleep(1)  # Give BIND time to be ready
-                    
-                    # Convert 0.0.0.0 to RedProxy container address for connection attempt
-                    # In Docker environments, 0.0.0.0 means "all interfaces" on RedProxy container
-                    # We need to connect to the RedProxy host, not localhost
-                    if bound_addr[0] == "0.0.0.0":
-                        connect_ip = self.env.redproxy_host  # Use RedProxy container's address
-                    else:
-                        connect_ip = bound_addr[0]
-                    connect_port = bound_addr[1]
-                    
-                    TestLogger.info(f"Client attempting to connect to {connect_ip}:{connect_port}")
-                    
-                    # Connect to the bound address - this should trigger the second SOCKS response
-                    reader, writer = await asyncio.open_connection(connect_ip, connect_port)
-                    
-                    # Send test data
-                    test_message = b"Hello from BIND client!\n"
-                    writer.write(test_message)
-                    await writer.drain()
-                    
-                    TestLogger.info("Client sent test message")
-                    
-                    # Keep connection alive briefly
-                    await asyncio.sleep(2)
-                    
-                    writer.close()
-                    await writer.wait_closed()
-                    TestLogger.info("Client connection closed")
-                    return True
-                    
-                except Exception as e:
-                    TestLogger.error(f"Client connection failed: {e}")
-                    return False
-            
-            # Step 3: Run the client connection simulation
-            client_task = asyncio.create_task(simulate_client_connection())
-            
-            try:
-                # Wait for client to complete with timeout
-                client_success = await asyncio.wait_for(client_task, timeout=10.0)
-                
-                if client_success:
-                    TestLogger.info("BIND connection flow test successful")
-                    self.add_test_result("bind_echo_test", True, "BIND connection flow completed successfully")
-                    return True
-                else:
-                    TestLogger.error("BIND connection flow failed")
-                    self.add_test_result("bind_echo_test", False, "Client connection to BIND address failed")
-                    return False
-                    
-            except asyncio.TimeoutError:
-                TestLogger.error("BIND connection flow timed out")
-                self.add_test_result("bind_echo_test", False, "BIND connection flow timed out")
-                return False
-            except Exception as e:
-                TestLogger.error(f"BIND connection flow error: {e}")
-                self.add_test_result("bind_echo_test", False, f"BIND flow error: {e}")
-                return False
-                
-        except Exception as e:
-            TestLogger.error(f"BIND test setup failed: {e}")
-            self.add_test_result("bind_echo_test", False, f"Test setup failed: {e}")
-            return False
-    
-    async def test_ipv6_bind_support(self) -> bool:
-        """Test BIND functionality with IPv6 addresses"""
-        TestLogger.test("Testing IPv6 BIND Support")
-        
-        try:
-            # First test IPv4 listener with IPv6 address request
-            TestLogger.info("Testing IPv6 BIND request on IPv4 listener")
-            socks_tester = SocksBindTester(self.env.redproxy_host, 1081)
-            
-            # Create IPv6 BIND request to IPv4 listener
-            TestLogger.info("Testing SOCKS5 BIND with IPv6 address on IPv4 listener")
-            reader, writer = await asyncio.open_connection(self.env.redproxy_host, 1081)
-            
-            try:
-                # SOCKS5 auth
-                writer.write(b'\x05\x01\x00')  # No auth
-                await writer.drain()
-                auth_response = await reader.read(2)
-                
-                if auth_response != b'\x05\x00':
-                    TestLogger.error("SOCKS5 auth failed for IPv6 test")
-                    return False
-                
-                # IPv6 BIND request - bind to ::1 (localhost IPv6)
-                bind_request = b'\x05\x02\x00\x04'  # Version, BIND, reserved, IPv6
-                bind_request += b'\x00' * 15 + b'\x01'  # ::1 address
-                bind_request += b'\x00\x00'  # Port 0 (system assigned)
-                
-                writer.write(bind_request)
-                await writer.drain()
-                
-                # Read response
-                response = await asyncio.wait_for(reader.read(1024), timeout=5.0)
-                
-                ipv4_result = False
-                if len(response) >= 4:
-                    if response[1] == 0x00:  # Success
-                        TestLogger.info("IPv6 BIND request accepted on IPv4 listener")
-                        ipv4_result = True
-                    else:
-                        TestLogger.info(f"IPv6 BIND request rejected on IPv4 listener (error: {response[1]})")
-                        ipv4_result = True  # Expected behavior
-                        
-            finally:
-                writer.close()
-                await writer.wait_closed()
-            
-            # Test 2: Try to connect to IPv6 listener (port 1084) if available
-            TestLogger.info("Testing connection to IPv6 listener (port 1084)")
-            ipv6_listener_result = False
-            try:
-                # Try to connect to IPv6 listener
-                reader_v6, writer_v6 = await asyncio.wait_for(
-                    asyncio.open_connection(self.env.redproxy_host, 1084), 
-                    timeout=3.0
-                )
-                
-                TestLogger.info("Successfully connected to IPv6 listener")
-                
-                # Test basic BIND on IPv6 listener  
-                writer_v6.write(b'\x05\x01\x00')  # No auth
-                await writer_v6.drain()
-                auth_response_v6 = await reader_v6.read(2)
-                
-                if auth_response_v6 == b'\x05\x00':
-                    # IPv4 BIND request on IPv6 listener
-                    bind_request_v6 = b'\x05\x02\x00\x01'  # Version, BIND, reserved, IPv4
-                    bind_request_v6 += b'\x7f\x00\x00\x01'  # 127.0.0.1
-                    bind_request_v6 += b'\x00\x00'  # Port 0
-                    
-                    writer_v6.write(bind_request_v6)
-                    await writer_v6.drain()
-                    
-                    response_v6 = await asyncio.wait_for(reader_v6.read(1024), timeout=5.0)
-                    
-                    if len(response_v6) >= 4 and response_v6[1] == 0x00:
-                        TestLogger.info("BIND successful on IPv6 listener")
-                        ipv6_listener_result = True
-                
-                writer_v6.close()
-                await writer_v6.wait_closed()
-                
-            except (asyncio.TimeoutError, ConnectionRefusedError):
-                TestLogger.info("IPv6 listener not available or not accessible")
-                ipv6_listener_result = True  # Not a failure if IPv6 isn't available
-            except Exception as e:
-                TestLogger.warn(f"IPv6 listener test failed: {e}")
-                ipv6_listener_result = False
-                
-            # Evaluate results
-            if ipv4_result and ipv6_listener_result:
-                self.add_test_result("ipv6_bind_test", True, "IPv6 BIND tests completed successfully")
-                return True
-            else:
-                self.add_test_result("ipv6_bind_test", False, f"IPv6 tests: IPv4={ipv4_result}, IPv6={ipv6_listener_result}")
-                return False
-                
-        except Exception as e:
-            TestLogger.error(f"IPv6 BIND test failed: {e}")
-            self.add_test_result("ipv6_bind_test", False, f"IPv6 test error: {e}")
-            return False
-    
-    async def test_concurrent_bind_operations(self) -> bool:
-        """Test multiple concurrent BIND operations"""
-        TestLogger.test("Testing Concurrent BIND Operations")
-        
-        async def single_bind_operation(bind_id: int):
-            try:
-                TestLogger.info(f"Starting BIND operation {bind_id}")
-                socks_tester = SocksBindTester(self.env.redproxy_host, 1081)
-                success, bound_addr = await socks_tester.socks5_bind_request("0.0.0.0", 0)
-                
-                if success and bound_addr:
-                    TestLogger.info(f"BIND {bind_id} successful: {bound_addr[0]}:{bound_addr[1]}")
-                    await asyncio.sleep(2)  # Hold the BIND briefly
-                    return True
-                else:
-                    TestLogger.warn(f"BIND {bind_id} failed")
-                    return False
-                    
-            except Exception as e:
-                TestLogger.error(f"BIND {bind_id} error: {e}")
-                return False
-        
-        try:
-            # Start 3 concurrent BIND operations
-            TestLogger.info("Starting 3 concurrent BIND operations")
-            tasks = [
-                asyncio.create_task(single_bind_operation(i))
-                for i in range(3)
-            ]
-            
-            # Wait for all to complete
-            results = await asyncio.gather(*tasks, return_exceptions=True)
-            
-            successful_binds = sum(1 for r in results if r is True)
-            TestLogger.info(f"Concurrent BIND results: {successful_binds}/3 successful")
-            
-            # At least 2 should succeed (allowing for some resource constraints)
-            if successful_binds >= 2:
-                self.add_test_result("concurrent_bind_test", True, f"{successful_binds}/3 concurrent BINDs successful")
-                return True
-            else:
-                self.add_test_result("concurrent_bind_test", False, f"Only {successful_binds}/3 concurrent BINDs successful")
-                return False
-                
-        except Exception as e:
-            TestLogger.error(f"Concurrent BIND test failed: {e}")
-            self.add_test_result("concurrent_bind_test", False, f"Concurrent test error: {e}")
-            return False
-    
-    async def test_bind_error_conditions(self) -> bool:
-        """Test BIND error handling and edge cases"""
-        TestLogger.test("Testing BIND Error Conditions")
-        
-        success_count = 0
-        total_tests = 0
-        
-        # Test 1: BIND to privileged port 
-        # Note: In Docker containers, ip_unprivileged_port_start is often set to 0,
-        # meaning all ports are unprivileged. This is different from host systems.
-        try:
-            total_tests += 1
-            TestLogger.info("Testing BIND to privileged port (22)")
-            
-            # First, check if we're in a container environment where this test makes sense
-            import os
-            container_detected = os.path.exists('/.dockerenv') or os.environ.get('container') == 'docker'
-            
-            socks_tester = SocksBindTester(self.env.redproxy_host, 1081)
-            success, bound_addr = await socks_tester.socks5_bind_request("127.0.0.1", 22)
-            
-            if container_detected:
-                # In containers, privileged ports are often allowed for non-root users
-                TestLogger.info(f"Container environment detected: BIND to port 22 {'succeeded' if success else 'failed'}")
-                success_count += 1  # Either outcome is acceptable in containers
-            else:
-                # On host systems, this should fail for non-root
-                if not success:
-                    TestLogger.info("BIND to privileged port correctly rejected on host system")
-                    success_count += 1
-                else:
-                    TestLogger.warn("BIND to privileged port unexpectedly succeeded on host system")
-                
-        except Exception as e:
-            TestLogger.info(f"BIND to privileged port failed as expected: {e}")
-            success_count += 1
-        
-        # Test 2: BIND with invalid address format
-        try:
-            total_tests += 1
-            TestLogger.info("Testing BIND with invalid address")
-            socks_tester = SocksBindTester(self.env.redproxy_host, 1081)
-            
-            # Try to make a malformed BIND request
-            reader, writer = await asyncio.open_connection(self.env.redproxy_host, 1081)
-            
-            try:
-                # SOCKS5 auth
-                writer.write(b'\x05\x01\x00')
-                await writer.drain()
-                await reader.read(2)
-                
-                # Malformed BIND request (invalid address type)
-                writer.write(b'\x05\x02\x00\xFF')  # Invalid address type 0xFF
-                await writer.drain()
-                
-                response = await asyncio.wait_for(reader.read(10), timeout=3.0)
-                if len(response) >= 2 and response[1] != 0x00:  # Error response
-                    TestLogger.info("Invalid BIND request correctly rejected")
-                    success_count += 1
-                    
-            finally:
-                writer.close()
-                await writer.wait_closed()
-                
-        except Exception as e:
-            TestLogger.info(f"Invalid BIND request handled: {e}")
-            success_count += 1
-        
-        # Test 3: Rapid BIND/unbind cycles
-        try:
-            total_tests += 1
-            TestLogger.info("Testing rapid BIND/unbind cycles")
-            
-            rapid_success = 0
-            for i in range(5):
-                socks_tester = SocksBindTester(self.env.redproxy_host, 1081)
-                success, bound_addr = await socks_tester.socks5_bind_request("0.0.0.0", 0)
-                if success:
-                    rapid_success += 1
-                # Connection automatically closes, releasing the BIND
-                await asyncio.sleep(0.1)
-            
-            if rapid_success >= 4:  # Allow for some failures
-                TestLogger.info(f"Rapid BIND cycles: {rapid_success}/5 successful")
-                success_count += 1
-            else:
-                TestLogger.warn(f"Rapid BIND cycles: only {rapid_success}/5 successful")
-                
-        except Exception as e:
-            TestLogger.error(f"Rapid BIND cycle test failed: {e}")
-        
-        # Test 4: BIND to port already in use (more reliable than privileged port test)
-        try:
-            total_tests += 1
-            TestLogger.info("Testing BIND to port already in use")
-            
-            # Try to BIND to the SOCKS port itself (should fail)
-            socks_tester = SocksBindTester(self.env.redproxy_host, 1081)
-            success, bound_addr = await socks_tester.socks5_bind_request("127.0.0.1", 1081)
-            
-            if not success:
-                TestLogger.info("BIND to occupied port correctly rejected")
-                success_count += 1
-            else:
-                TestLogger.warn("BIND to occupied port unexpectedly succeeded")
-                # This might succeed if RedProxy uses SO_REUSEADDR, which is valid behavior
-                success_count += 1
-                
-        except Exception as e:
-            TestLogger.info(f"BIND to occupied port handled: {e}")
-            success_count += 1
-        
-        # Evaluate overall success
-        TestLogger.info(f"Error condition tests: {success_count}/{total_tests} passed")
-        
-        if success_count >= total_tests - 1:  # Allow one failure
-            self.add_test_result("bind_error_conditions", True, f"{success_count}/{total_tests} error condition tests passed")
-            return True
-        else:
-            self.add_test_result("bind_error_conditions", False, f"Only {success_count}/{total_tests} error condition tests passed")
-            return False
-    
-    async def test_override_bind_address(self) -> bool:
-        """Test override_bind_address functionality for NAT scenarios"""
-        TestLogger.test("Testing override_bind_address NAT functionality")
-        
-        try:
-            # Test 1: Normal BIND (baseline) - port 1081 uses regular direct connector
-            TestLogger.info("Phase 1: Testing normal BIND response (baseline)")
-            socks_tester_normal = SocksBindTester(self.env.redproxy_host, 1081)
-            success_normal, bound_addr_normal = await socks_tester_normal.socks5_bind_request("0.0.0.0", 0)
-            
-            if not success_normal or not bound_addr_normal:
-                self.add_test_result("override_bind_address", False, "Baseline BIND test failed")
-                return False
-            
-            TestLogger.info(f"Normal BIND response: {bound_addr_normal[0]}:{bound_addr_normal[1]}")
-            
-            # Test 2: Override BIND - port 1085 uses direct-with-override connector
-            TestLogger.info("Phase 2: Testing override BIND response (port 1085)")
-            socks_tester_override = SocksBindTester(self.env.redproxy_host, 1085)
-            success_override, bound_addr_override = await socks_tester_override.socks5_bind_request("0.0.0.0", 0)
-            
-            if not success_override or not bound_addr_override:
-                TestLogger.error("Override BIND test failed")
-                self.add_test_result("override_bind_address", False, "Override BIND request failed")
-                return False
-            
-            TestLogger.info(f"Override BIND response: {bound_addr_override[0]}:{bound_addr_override[1]}")
-            
-            # Test 3: Verify the override worked
-            expected_override_ip = "192.168.1.100"
-            
-            if bound_addr_override[0] == expected_override_ip:
-                TestLogger.info(f"SUCCESS: Override address correctly returned as {expected_override_ip}")
-                TestLogger.info(f"Normal address: {bound_addr_normal[0]}, Override address: {bound_addr_override[0]}")
-                self.add_test_result("override_bind_address", True, 
-                                   f"Override successful: normal={bound_addr_normal[0]}, override={bound_addr_override[0]}")
-                return True
-            else:
-                TestLogger.error(f"FAILED: Expected override address {expected_override_ip}, got {bound_addr_override[0]}")
-                self.add_test_result("override_bind_address", False, 
-                                   f"Override failed: expected {expected_override_ip}, got {bound_addr_override[0]}")
-                return False
-            
-        except Exception as e:
-            TestLogger.error(f"override_bind_address test failed: {e}")
-            self.add_test_result("override_bind_address", False, f"Test error: {e}")
-            return False
-
-    async def test_bind_timeout(self) -> bool:
-        """Test BIND timeout functionality to ensure BIND operations don't run indefinitely"""
-        TestLogger.test("Testing BIND Timeout Functionality")
-        
-        try:
-            # Test 1: Create a BIND connection that should timeout
-            TestLogger.info("Phase 1: Testing BIND timeout behavior")
-            
-            # Use the timeout test listener (port 1086) with very short idle timeout
-            socks_tester = SocksBindTester(self.env.redproxy_host, 1086)
-            
-            # Set up BIND request - this will succeed initially
-            success, bound_addr = await socks_tester.socks5_bind_request("0.0.0.0", 0)
-            
-            if not success or not bound_addr:
-                TestLogger.error("Initial BIND setup failed")
-                self.add_test_result("bind_timeout_test", False, "Initial BIND setup failed")
-                return False
-            
-            TestLogger.info(f"BIND established on {bound_addr[0]}:{bound_addr[1]} - waiting for timeout...")
-            
-            # Test 2: Wait longer than the configured timeout (3 seconds) to see if the task times out
-            # With the shortened timeout, the BIND task should terminate due to timeout
-            
-            start_time = time.time()
-            connection_timeout = False
-            
-            try:
-                # Try to connect to the bound address after the timeout period
-                # In a normal scenario, we'd expect this to either:
-                # 1. Connect successfully (if no timeout implemented)  
-                # 2. Connection refused (if timeout properly terminates the BIND task)
-                
-                await asyncio.sleep(4.0)  # Wait longer than the 3s timeout
-                
-                TestLogger.info("Attempting connection to BIND address after timeout period")
-                
-                # Try to connect - this should fail if timeout worked properly
-                try:
-                    reader, writer = await asyncio.wait_for(
-                        asyncio.open_connection(self.env.redproxy_host, bound_addr[1]), 
-                        timeout=2.0
-                    )
-                    
-                    # If we got here, the connection succeeded, meaning the timeout didn't work
-                    TestLogger.warn("Connection to BIND address succeeded after timeout period - timeout may not be working")
-                    writer.close()
-                    await writer.wait_closed()
-                    connection_timeout = False
-                    
-                except (ConnectionRefusedError, OSError, asyncio.TimeoutError):
-                    # Connection failed - this is what we expect if timeout worked
-                    TestLogger.info("Connection to BIND address failed after timeout - timeout appears to be working")
-                    connection_timeout = True
-                    
-            except Exception as e:
-                TestLogger.error(f"Error during timeout test: {e}")
-                connection_timeout = True  # Assume timeout worked if there was an error
-            
-            elapsed_time = time.time() - start_time
-            TestLogger.info(f"Total test time: {elapsed_time:.1f} seconds")
-            
-            # Test 3: Verify behavior with fresh connection
-            TestLogger.info("Phase 2: Verifying timeout behavior with fresh connection")
-            
-            # Try a new BIND operation to ensure the system is still responsive
-            socks_tester_fresh = SocksBindTester(self.env.redproxy_host, 1081)  # Use normal port
-            success_fresh, bound_addr_fresh = await socks_tester_fresh.socks5_bind_request("0.0.0.0", 0)
-            
-            if success_fresh and bound_addr_fresh:
-                TestLogger.info(f"Fresh BIND successful after timeout test: {bound_addr_fresh[0]}:{bound_addr_fresh[1]}")
-            else:
-                TestLogger.warn("Fresh BIND failed - system may be in bad state")
-            
-            # Evaluate results
-            # Success criteria:
-            # 1. Initial BIND setup worked
-            # 2. After timeout period, connection to BIND address fails (indicating timeout worked)
-            # 3. Fresh BIND operations still work (system not broken)
-            
-            success_criteria_met = success and connection_timeout and success_fresh
-            
-            if success_criteria_met:
-                TestLogger.info("SUCCESS: BIND timeout functionality working correctly")
-                self.add_test_result("bind_timeout_test", True, 
-                                   f"Timeout test passed: initial BIND OK, timeout after {elapsed_time:.1f}s, fresh BIND OK")
-                return True
-            else:
-                TestLogger.warn("PARTIAL: BIND timeout test showed mixed results")
-                # Consider this a success if at least the basic functionality works
-                # Timeout behavior can be hard to test reliably in containerized environments
-                if success and success_fresh:
-                    TestLogger.info("Basic BIND functionality confirmed, timeout behavior uncertain")
-                    self.add_test_result("bind_timeout_test", True, 
-                                       f"Basic BIND OK, timeout behavior uncertain (elapsed: {elapsed_time:.1f}s)")
-                    return True
-                else:
-                    self.add_test_result("bind_timeout_test", False, 
-                                       f"BIND timeout test failed: initial={success}, timeout={connection_timeout}, fresh={success_fresh}")
-                    return False
-            
-        except Exception as e:
-            TestLogger.error(f"BIND timeout test failed: {e}")
-            self.add_test_result("bind_timeout_test", False, f"Test error: {e}")
-            return False
-
-    async def run_all_tests(self) -> bool:
-        """Run all BIND tests"""
-        TestLogger.info(f"{Colors.BLUE}=== Starting BIND Functionality Test Suite ==={Colors.NC}")
-        
-        tests = [
-            ("SOCKS Listener BIND", self.test_socks_listener_bind_support),
-            ("SOCKS Connector BIND", self.test_socks_connector_bind_functionality),
-            ("DirectConnector BIND", self.test_direct_connector_bind_functionality),
-            ("BIND Connection Flow", self.test_bind_with_echo_server),
-            ("IPv6 BIND Support", self.test_ipv6_bind_support),
-            ("Concurrent BIND Operations", self.test_concurrent_bind_operations),
-            ("BIND Error Conditions", self.test_bind_error_conditions),
-            ("Override BIND Address", self.test_override_bind_address),
-            ("BIND Timeout", self.test_bind_timeout),
-        ]
-        
-        passed = 0
-        total = len(tests)
-        
-        for test_name, test_func in tests:
-            TestLogger.info(f"\n{Colors.YELLOW}--- {test_name} ---{Colors.NC}")
-            try:
-                if await test_func():
-                    TestLogger.info(f"{Colors.GREEN}✓ {test_name} PASSED{Colors.NC}")
-                    passed += 1
-                else:
-                    TestLogger.error(f"{Colors.RED}✗ {test_name} FAILED{Colors.NC}")
-            except Exception as e:
-                TestLogger.error(f"{Colors.RED}✗ {test_name} ERROR: {e}{Colors.NC}")
-                self.add_test_result(test_name.lower().replace(' ', '_'), False, f"Exception: {e}")
-        
-        # Generate report
-        success_rate = (passed / total) * 100
-        TestLogger.info(f"\n{Colors.BLUE}=== BIND Test Suite Results ==={Colors.NC}")
-        TestLogger.info(f"Tests passed: {passed}/{total} ({success_rate:.1f}%)")
-        
-        # Finalize suite and generate reports
-        self.reporter.finalize_suite(self.suite)
-        json_path = self.reporter.save_json_report("bind_report.json")
-        html_path = self.reporter.save_html_report("bind_report.html")
-        
-        if passed == total:
-            TestLogger.info(f"{Colors.GREEN}All BIND tests PASSED!{Colors.NC}")
-            TestLogger.info(f"Reports saved: {json_path}, {html_path}")
-            return True
-        else:
-            TestLogger.error(f"{Colors.RED}Some BIND tests FAILED!{Colors.NC}")
-            TestLogger.info(f"Reports saved: {json_path}, {html_path}")
-            return False
-
-
-async def main():
-    """Main test execution"""
-    try:
-        # Setup test environment
-        env = setup_test_environment()
-        
-        # Wait for basic services
-        TestLogger.info("Waiting for RedProxy services...")
-        if not await wait_for_service(env.redproxy_host, env.redproxy_socks_port):
-            TestLogger.error("RedProxy SOCKS service not available")
-            return False
-        
-        # Run BIND test suite
-        test_suite = BindTestSuite(env)
-        success = await test_suite.run_all_tests()
-        
-        return success
-        
-    except KeyboardInterrupt:
-        TestLogger.warn("Test interrupted by user")
-        return False
-    except Exception as e:
-        TestLogger.error(f"Test suite failed: {e}")
-        return False
-
-
-if __name__ == "__main__":
-    success = asyncio.run(main())
-    exit(0 if success else 1)
